refactored accounts file structure:
