+++ routes/api/accounts/index.js
@@ -1,22 +1,14 @@
 const { Router } = require('express');
 const router = Router();
 
-router.get('/:id', async (req, res) => {
-  const db = req.app.get('db');
-  const { id } = req.params;
-  const user = await db.User.findById(id);
-  if (!user) {
-    return res.status(404).json({
-      status: 404,
-      statusText: 'Not Found',
-    });
-  }
-  res.status(200).json({
-    status: 200,
-    statusText: 'OK',
-    result: { user },
-  });
-});
+router.use(
+  '/:id',
+  (req, res, next) => {
+    req.user = { id: req.params.id };
+    next();
+  },
+  require('./-id'),
+);
 
 router.post('/', async (req, res) => {
   const db = req.app.get('db');
